% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{dbmerge}
\alias{dbmerge}
\title{Merge columns from multiple dataframes}
\usage{
dbmerge(data1, data2, key, selection, selection2 = NULL, output = FALSE)
}
\arguments{
\item{data1}{A data.frame, the data frame to which new columns will be added}

\item{data2}{A data.frame, the data frame from which columns will be extracted}

\item{key}{Character vector with names of the columns which contain the keys}

\item{selection}{Character vector with names of columns which will be extracted
from data2}

\item{selection2}{Optional character vector with names of columns from data1
that want to be kept, all others will be removed}

\item{output}{Logical, should a warning message be output if key matches multiple entries}
}
\value{
A new data.frame object
}
\description{
This function uses a common column (i.e. key in relational databases) to
add columns from one data frame to another.
}
\details{
This function takes two data.frames and copies the entries from one data.frame
to another as matched by one or multiple "key" columns found in both data.frames. This
function can be used to combine the relevant data from multiple tables from a relational
database. NOTE: if a key matches two entries in data2, the data from the first entry
is used and a warning message is output.
}
